Enable middleware in TestScript and refactor BeginSkill tests
