+++ libraries/Microsoft.Bot.Builder.Dialogs.Adaptive.Testing/TestScript.cs
@@ -131,9 +131,10 @@ namespace Microsoft.Bot.Builder.Dialogs.Adaptive.Testing
         /// </summary>
         /// <param name="resourceExplorer">Resource explorer to use.</param>
         /// <param name="testName">Name of test.</param>
+        /// <param name="middlweare">Middleware to add to the adapter.</param>
         /// <returns>Test adapter.</returns>
 #pragma warning disable CA1801 // Review unused parameters (excluding for now but consider removing the resourceExplorer parameter if it is not needed)
-        public TestAdapter DefaultTestAdapter(ResourceExplorer resourceExplorer, [CallerMemberName] string testName = null)
+        public TestAdapter DefaultTestAdapter(ResourceExplorer resourceExplorer, [CallerMemberName] string testName = null, IEnumerable<IMiddleware> middlweare = null)
 #pragma warning restore CA1801 // Review unused parameters
         {
             var storage = new MemoryStorage();
@@ -145,9 +146,15 @@ namespace Microsoft.Bot.Builder.Dialogs.Adaptive.Testing
                 .UseStorage(storage)
                 .UseBotState(userState, convoState)
                 .Use(new TranscriptLoggerMiddleware(new TraceTranscriptLogger(traceActivity: false)))
-                .Use(new SetTestOptionsMiddleware())
-                .Use(new SetSkillConversationIdFactoryBaseMiddleware())
-                .Use(new SetSkillBotFrameworkClientMiddleware());
+                .Use(new SetTestOptionsMiddleware());
+
+            if (middlweare != null)
+            {
+                foreach (var m in middlweare)
+                {
+                    adapter.Use(m);
+                }
+            }
 
             adapter.OnTurnError += (context, err) => context.SendActivityAsync(err.Message);
             return adapter;
@@ -162,12 +169,13 @@ namespace Microsoft.Bot.Builder.Dialogs.Adaptive.Testing
         /// <param name="testName">Name of the test.</param>
         /// <param name="callback">The bot logic.</param>
         /// <param name="adapter">optional test adapter.</param>
+        /// <param name="middlweare">Middleware to add to the adapter.</param>
         /// <returns>Runs the exchange between the user and the bot.</returns>
-        public async Task ExecuteAsync(ResourceExplorer resourceExplorer, [CallerMemberName] string testName = null, BotCallbackHandler callback = null, TestAdapter adapter = null)
+        public async Task ExecuteAsync(ResourceExplorer resourceExplorer, [CallerMemberName] string testName = null, BotCallbackHandler callback = null, TestAdapter adapter = null, IEnumerable<IMiddleware> middlweare = null)
         {
             if (adapter == null)
             {
-                adapter = DefaultTestAdapter(resourceExplorer, testName);
+                adapter = DefaultTestAdapter(resourceExplorer, testName, middlweare);
             }
 
             adapter.EnableTrace = EnableTrace;
